Slightly more advanced Rock, Paper, Scissors program
This "more advanced" program uses randint to assign the computer a choice. It will take user input, then compare the input to the computer's choice using conditional logic to determine a winner. The user input is made case-insensitive by using the lower() function to make the input all lowercase letters.